+++ src/QueryBuilder/AbstractDQLQueryBuilder.php
@@ -182,9 +182,9 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
 
     public function buildExpression(ExpressionInterface $expression, array &$params = []): string
     {
-        $builder = $this->queryBuilder->getExpressionBuilder($expression);
-        /** @psalm-suppress MixedMethodCall */
-        return (string) $builder->build($expression, $params);
+        return $this->queryBuilder
+            ->getExpressionBuilder($expression)
+            ->build($expression, $params);
     }
 
     public function buildFor(array $values): string
@@ -468,7 +468,7 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
         return count($conditions) === 1 ? $conditions[0] : new Condition\AndX(...$conditions);
     }
 
-    public function getExpressionBuilder(ExpressionInterface $expression): object
+    public function getExpressionBuilder(ExpressionInterface $expression): ExpressionBuilderInterface
     {
         $className = $expression::class;